Fixed deletion on null db file

+++ lib/sqlite/local_storage.dart
@@ -43,10 +43,8 @@ class LocalStorage {
     final docPath = (await getApplicationDocumentsDirectory()).path;
     final path = join(docPath, databasePath);
     File dbFile = File(path);
-    try {
+    if (await dbFile.exists()) {
       await dbFile.delete();
-    } catch (e) {
-      // TODO
     }
     _db = await _dbFactory.openDatabase(path);
     final tags = await _firestoreManager.getTags();
@@ -67,7 +65,9 @@ class LocalStorage {
     final docPath = (await getApplicationDocumentsDirectory()).path;
     final path = join(docPath, databasePath);
     File dbFile = File(path);
-    await dbFile.delete();
+    if (await dbFile.exists()) {
+      await dbFile.delete();
+    }
     _db = await _dbFactory.openDatabase(path);
     final tags = await _firestoreManager.getTags();
     _importTags(tags);
